Refactor code for better structure and readability
The DeepPartialObject<T> type has been moved out of the deepMerge.ts file and into the _recursiveMerge.ts file on its own. Additionally, functions are now accept objects instead of arrays.
